third_party/blink/web_tests/crypto/random-values-limits-expected.txt

Tests the limits of crypto.randomValues.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS 'crypto' in self is true
PASS 'getRandomValues' in self.crypto is true
PASS crypto.getRandomValues(almostTooLargeArray) did not throw exception.
PASS crypto.getRandomValues(tooLargeArray) threw exception QuotaExceededError: Failed to execute 'getRandomValues' on 'Crypto': The ArrayBufferView's byte length (65537) exceeds the number of bytes of entropy available via this API (65536)..
PASS successfullyParsed is true

TEST COMPLETE
